zdoom: Add legacy zdoom recipe
Adds a recipe for the (original) legacy version of zdoom that was
discontinued in 2016. This recipe should only be used if necessary to
support the oldest hardware, although be aware it appears to only do
software rendering to textures on accelerated platforms.

The licensing on this version of zdoom is a mess since it pulls in so
many components with difference licenses. I've purposely omitted the
LICENSE statement (so the recipe won't parse) until it's figured out...
If I don't end up needing this recipe, I may not bother.

This commit is mostly so I don't loose the work I did if I do need to
resurrect it.